Finished styling the notes page for the color and completed the color intergration in the notesviewmodel, new notesviewmodel and editnoteviewmodel.
Added a check to not always reorder the list when the page has loaded

Want to add functionality so that the notes don't have to constantly get removed inorder to add new items on the OnAppearing actions.

Changed the "transparent" note color to be the same as the note item background.